Move Modbus maps, model, and device into client.modbus
Relocate the datasheet maps, typed components, and HTTP-shaped adapter
under custom_components/qvantum/client/modbus. Keep thin re-exports at
the old module paths so the integration and tests stay unchanged.
