+++ custom_components/qvantum/modbus.py
@@ -1,179 +1,17 @@
-"""Modbus constants and mappings for the Qvantum Heat Pump Integration.
+"""Compatibility re-export; datasheet maps live in ``client.modbus.maps``."""
 
-These maps are the datasheet. ``modbus_model`` builds typed components from
-them, and ``modbus_device`` adapts decoded values to the HTTP-shaped payloads
-the rest of the integration already consumes.
-"""
+from .client.modbus.maps import (
+    MODBUS_HOLDING_REGISTER_MAP,
+    MODBUS_HOLDING_TO_SETTINGS_MAP,
+    MODBUS_IDENTITY_REGISTER_MAP,
+    MODBUS_INPUT_REGISTER_MAP,
+    RELAY_BIT_MAP,
+)
 
-# Modbus input register map using internal metric names as keys.
-# Format: internal_name -> (register_address, data_type, scale)
-MODBUS_INPUT_REGISTER_MAP = {
-    "bt1": (0, "int16", 0.1),
-    "bt2": (2, "int16", 0.1),
-    "bt4": (4, "int16", 0.1),
-    "bt10": (5, "int16", 0.1),
-    "bt11": (6, "int16", 0.1),
-    "bt12": (7, "int16", 0.1),
-    "bt13": (8, "int16", 0.1),
-    "bt14": (9, "int16", 0.1),
-    "bt15": (10, "int16", 0.1),
-    "bt20": (11, "int16", 0.1),
-    "bt21": (12, "int16", 0.1),
-    "bt22": (13, "int16", 0.1),
-    "bt23": (14, "int16", 0.1),
-    "bt30": (16, "int16", 0.1),
-    "bt31": (17, "int16", 0.1),
-    "bt33": (18, "int16", 0.1),
-    "bt34": (19, "int16", 0.1),
-    "btx": (21, "int16", 0.1),
-    "bf1_l_min": (26, "uint16", 0.01),
-    "bf1_rpm": (27, "uint16", 0.1),
-    "gp1_speed": (28, "uint16", 0.1),
-    "gp2_speed": (29, "uint16", 1.0),
-    "fanrpm": (30, "uint16", 1.0),
-    "compressormeasuredspeed": (31, "uint16", 1.0),
-    "relays_bitmask": (33, "uint16", 1.0),
-    "degree_minute": (34, "int16", 0.1),
-    "cal_heat_temp": (35, "int16", 0.1),
-    "unit_state": (40, "uint16", 1.0),
-    "hp_status": (41, "uint16", 1.0),
-    "heatingreleased": (42, "uint16", 1.0),
-    "coolingreleased": (43, "uint16", 1.0),
-    "compressorreleased": (44, "uint16", 1.0),
-    "additionreleased": (45, "uint16", 1.0),
-    "heatingdemand": (50, "uint16", 1.0),
-    "coolingdemand": (51, "uint16", 1.0),
-    "additiondemand": (52, "uint16", 1.0),
-    "additiondhwdemand": (53, "uint16", 0.1),
-    "dhwdemand": (54, "uint16", 1.0),
-    "heating_prioritytimeleft": (63, "uint16", 1.0),
-    "cooling_prioritytimeleft": (64, "uint16", 1.0),
-    "dhw_prioritytimeleft": (65, "uint16", 1.0),
-    "time_to_defrost": (67, "uint16", 1.0),
-    "compressor_state": (70, "uint16", 1.0),
-    "compressor_blocked": (71, "uint16", 1.0),
-    "compressor_blocked_sec": (72, "uint16", 1.0),
-    "qn8position": (76, "int16", 1.0),
-    "freeze_protection_active": (87, "uint16", 1.0),
-    "ventilation_filter_time_left": (91, "uint16", 1.0),
-    "compressor_power": (93, "uint16", 1.0),
-    "compressor_mwh": (95, "uint16", 1.0),
-    "compressor_kwh": (96, "uint16", 0.1),
-    "additional_mwh": (97, "uint16", 1.0),
-    "additional_kwh": (98, "uint16", 0.1),
-    "heating_mwh": (99, "uint16", 1.0),
-    "heating_kwh": (100, "uint16", 0.1),
-    "cooling_mwh": (101, "uint16", 1.0),
-    "cooling_kwh": (102, "uint16", 0.1),
-    "dhw_mwh": (103, "uint16", 1.0),
-    "dhw_kwh": (104, "uint16", 0.1),
-    "smart_dhw_mode": (161, "uint16", 1.0),
-    "smart_dhw_control_status": (162, "uint16", 1.0),
-    "enable_sc_dhw": (163, "uint16", 1.0),
-    "enable_sc_sh": (164, "uint16", 1.0),
-    "wifi_connected": (168, "uint16", 1.0),
-    "cloud_connected": (169, "uint16", 1.0),
-    "vacation_mode": (170, "uint16", 1.0),
-}
-
-# Device identity (QAD EN 2609-AXC input 180-193). Kept off the metrics map so
-# a refused identity block cannot fail the 0-104 / 161-170 poll.
-MODBUS_IDENTITY_REGISTER_MAP = {
-    "serial_1": (180, "uint16", 1.0),
-    "serial_2": (181, "uint16", 1.0),
-    "serial_3": (182, "uint16", 1.0),
-    "serial_4": (183, "uint16", 1.0),
-    "serial_5": (184, "uint16", 1.0),
-    "ip_1": (186, "uint16", 1.0),
-    "ip_2": (187, "uint16", 1.0),
-    "ip_3": (188, "uint16", 1.0),
-    "ip_4": (189, "uint16", 1.0),
-    "version_major": (191, "uint16", 1.0),
-    "version_minor": (192, "uint16", 1.0),
-    "version_patch": (193, "uint16", 1.0),
-}
-
-MODBUS_HOLDING_REGISTER_MAP = {
-    "unit_on_off": (0, "uint16", 1.0),
-    "operation_mode": (1, "uint16", 1.0),
-    "allow_heating": (2, "uint16", 1.0),
-    "allow_cooling": (3, "uint16", 1.0),
-    "allow_addition": (4, "uint16", 1.0),
-    "allow_dhw": (5, "uint16", 1.0),
-    "time_between_modes": (6, "uint16", 1.0),
-    "allow_addition_temp": (7, "int16", 1.0),
-    "filtertime_outdoor": (8, "uint16", 1.0),
-    "use_operation_sensor": (9, "uint16", 1.0),
-    "desired_indoor_temp": (12, "int16", 0.1),
-    "room_compensation": (13, "uint16", 0.1),
-    "room_temp_external": (14, "int16", 0.1),
-    "heating_offset": (15, "int16", 1.0),
-    "stop_heating": (18, "int16", 1.0),
-    "max_heating_supply": (19, "int16", 1.0),
-    "min_heating_supply": (20, "uint16", 1.0),
-    "curve_type_heating": (22, "uint16", 1.0),
-    "temp_compensation_curve": (23, "uint16", 1.0),
-    "cooling_offset": (36, "int16", 1.0),
-    "start_cooling_temp": (38, "int16", 1.0),
-    "dew_point_protection": (39, "uint16", 1.0),
-    "min_cooling_supply": (40, "int16", 1.0),
-    "dhw_mode": (53, "uint16", 1.0),
-    "dhw_start_normal": (56, "uint16", 1.0),
-    "dhw_stop_normal": (57, "uint16", 1.0),
-    "dhw_start_extra": (58, "uint16", 1.0),
-    "dhw_stop_extra": (59, "uint16", 1.0),
-    "dhw_outlet_temp": (60, "uint16", 1.0),
-    "dhw_uninterrupted_cooling": (61, "uint16", 1.0),
-    "pump_speed_heating": (63, "uint16", 1.0),
-    "pump_speed_cooling": (64, "uint16", 1.0),
-    "pump_speed_dhw": (65, "uint16", 1.0),
-    "pump_idle_speed": (66, "uint16", 1.0),
-    "ventilation_state": (68, "uint16", 1.0),
-    "fan_speed_reduced": (69, "uint16", 1.0),
-    "fan_speed_normal": (70, "uint16", 1.0),
-    "fan_speed_extra": (71, "uint16", 1.0),
-    "compressor_fan_speed": (72, "uint16", 1.0),
-    "heating_priority_time": (73, "uint16", 1.0),
-    "cooling_priority_time": (74, "uint16", 1.0),
-    "dhw_priority_time": (75, "uint16", 1.0),
-    "bt12_mounted": (83, "uint16", 1.0),
-    "qs_unit_connected": (84, "uint16", 1.0),
-    "sg_enabled": (88, "uint16", 1.0),
-}
-
-RELAY_BIT_MAP = {
-    "picpin_relay_heat_l1": 0,
-    "picpin_relay_heat_l2": 1,
-    "picpin_relay_heat_l3": 2,
-    "picpin_relay_gp10": 3,
-    "picpin_relay_qm10": 4,
-    "picpin_relay_qn8_1": 5,
-    "picpin_relay_qn8_2": 6,
-    "picpin_relay_gp3": 7,
-    "picpin_relay_pump": 8,
-    "picpin_relay_ha12": 9,
-}
-
-# Map from Modbus holding register keys to the canonical settings names used by
-# the API/settings layer. Also used in reverse (settings name -> holding key) by
-# write_holding_register_for_metric() to resolve a metric key to its register address.
-MODBUS_HOLDING_TO_SETTINGS_MAP = {
-    "start_cooling_temp": "start_cooling_temp",
-    "dhw_start_normal": "tap_water_start",
-    "dhw_stop_normal": "tap_water_stop",
-    "dhw_mode": "extra_tap_water",
-    "room_compensation": "room_comp_factor",
-    "desired_indoor_temp": "indoor_temperature_target",
-    "heating_offset": "indoor_temperature_offset",
-    "use_operation_sensor": "sensor_mode",
-    "fan_speed_normal": "fan_normal",
-    "fan_speed_reduced": "fan_speed_2",
-    "operation_mode": "op_mode",
-    "allow_dhw": "op_man_dhw",
-    "allow_addition": "op_man_addition",
-    "allow_heating": "man_mode",
-    "ventilation_state": "fanspeedselector",
-    "dhw_stop_extra": "dhw_stop_extra",
-    "room_temp_external": "room_temp_external",
-    "stop_heating": "stop_heating",
-}
+__all__ = [
+    "MODBUS_HOLDING_REGISTER_MAP",
+    "MODBUS_HOLDING_TO_SETTINGS_MAP",
+    "MODBUS_IDENTITY_REGISTER_MAP",
+    "MODBUS_INPUT_REGISTER_MAP",
+    "RELAY_BIT_MAP",
+]
